Handle commits in conf.json and in templates
Right now, klp-build is able to gather backported commits by using the
commit subject of an upstream commit.

Current it required a KLP_KERNEL_SOURCE env var to be defined. klp-build
setup-bsc will got to this dir, for each codestream base kernel and look
for the same commit subject found in the upstream fix. These commits are
stored in conf.json file to be later used by klp-build create-lp
subcommand.

There are still some minor formatting issues, but these can be solved
when creating the livepatch.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>