klp-build: Add get-patches subcommand
This new subcommand will be responsible for getting the backported
commits of each CVE branch, related to the upstream commits passed as
arguments.

This was being doing for the setup subcommand as well, but it can be
used to inspect which files need to be touched even before the setup
phase.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>