utils.py: Change --file-funcs to accept a regex per codestream
Until now --file-funcs was expecting a kernel family as the first
argument. But we can have a kernel family with different patches
backported, resulting in very different kernels for the same base kernel
version.

Using the new approach results in a more fine grained selection of which
codestreams will patch which files/functions.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>