utils.py: Append all matching regex to a codestream
This can lead us to the following use case:

Let's say that only a file changes between all the code streams.
Currently we need to specify all files under the same codestream in
order to have them merged later in the process, like:

--file-funcs 12.5 a.c func1 func2 --file-funcs 12.5 b.c func3 func4
--file-funcs 15.3 a.c func1 func2 --file-funcs 15.3 b.c funcX func4

in this case, for 12.5 and 15.3 the functions of file a.c are the same.
The new method allows the user to specify only what is different between
the codestreams, like:

--file-funcs 12.5 b.c func3 func4 --file-funcs 15.3 b.c funX func4
--file-funcs ".*" a.c func1 func2

In livepatches that touch a lot of different codestreams and have
different function names depending on the codestream, the --file-funcs
can get huge. This commit makes it easier to manage such changes

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>