Add Config.filter_cs
This new method will execute the filtering that was duplicated between
util and ccp modules, also informing the user which codestreams are
being avoided.

Also, move up the get_patched_cs and get_commits since it was being
called before self.codestreams was populated. This only happened when
executed by the first time, but better to avoid such case.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>