Move symbol checking code to config and expand it to other arches
Now that the IBS module is able to download kernel-default package for
s390 and ppc64le, when supported by the codestream, klp-build can check
if the externalized symbols are found in the objects to be patches in
different arch objects.

The checking is done after klp-ccp finishes, since it builds the list of
externalized symbols.

Also, cache the output of the nm command by object path. The path
differs from codestream and architecture, so it helps in the case of a
repeated check for different symbols of a specific codestream/arch.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>