Implement check for livepatch module dependencies
A livepatch shouldn't have any dependencies. It should rely only on
symbols exported by vmlinux. So the code checks if the livepatch depends
on other modules, and if yes, it tries to find symbols that are not
defined in vmlinux, meaning that it 's defined in a module.

It throws a warning to the developer showing which functions couldn't
be found.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>