ccp.py: Try harder to group codestreams with expanded macros
When externalized symbols are used on macros, these are expanded, and
these can contain the full path of the affected codestream, making the
grouping process not useful. This change helps a bit to group similar
codestreams, at least making the user to check differences on less
codestreams.

In this case, managed to reduce the creation of different codestreams
from 22 (all codestreams were different) to 14.

Grouping codestreams that share the same content and files:
	15.4rtu1
	15.4rtu2
	15.4rtu3
	15.4rtu5
	15.4rtu6
	15.4rtu7
	15.4rtu8
	15.4rtu9
	15.4u4
	15.4u5
	15.4u6
	15.4u7
	15.4u8
	15.5u0
	15.5u1
	15.4rtu11
	15.4rtu13
	15.4u15
	15.4u16
	15.4u18
	15.5u2
	15.5u3
	15.5u5

Grouping codestreams that share the same content and files:
        15.4rtu1
        15.4rtu2-5
        15.4rtu6-7
        15.4rtu8-9
        15.4u4-5
        15.4u6
        15.4u7-8
        15.5u0
        15.5u1
        15.4rtu11
        15.4rtu13
        15.4u15-16
        15.4u18
        15.5u2-3
        15.5u5

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>