Introcuce run-{ccp,ce} --apply-patches
For now it tries to apply patches from $KGR_LP/livepatch/$bsc/fixes, and
applies the related patches.

For example, if we are executing run-ccp for 12.5, it will try first to
apply patches from fixes/12.5, and if it fails, try also fixes/cve-4.12.

For 15.3 ones, try first to fixes/15.3rt, then fixes/15.3 and then
fixes/cve-5.3.

For 15.4 onwards, we don't have cve-5.3, since the base kernel is 5.14.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>