config.py: Various fixes
Flush messages before running quilt commands, otherwise the message will
be shown after the quilt commands.

Fix the directories of patches to try out.

Reverse sort the patches, so we import them in the reverse order that
they will be applied. Otherwise quilt will apply patches 3, 2 and 1,
instead of 1, 2 and 3.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>