Remove work-dir variable
Let's remove this recently added argument. It doesn't add up to the LP
creation to specify a different output based on --kdir. Now we always
check if kdir was set on setup phase, and then choose different output
based on it. So now only require the work_dir to be specified on env
var.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>