check_symbol_archs: Fix arguments
The flags were inverted: When the setup is running we want it to check
the symbol on all architecures, including the host one. On extraction,
we can safely skip the host architecture, since we have rely on the
backend to make the right assumptions regarding the host architecture
w.r.t public symbols usage.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>