setup: Fix issue when erasing conf for a specific codestream
At setup_codestreams step, we assign the same dict to all codestreams
specified on --codestreams. After we check if the config is enabled for
the codestream, if the config is available in all architectures, we just
remove the config, since it's available everywhere. But as all
codestream shared the same dict, it would erase the config entry for all
of them, making it fail to detect if the config is enabled for the next
codestream in the loop.

This can be solved by copying the dict for each codestream.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>