ibs.py: Do not create a second time the ".config"
Once the kernel and related binaries get extracted so does the
".config", which gets copied twice to its respective kernel
src directory. Only once is good enough.

This commit also fixes a bug introduced in fb40c1c5, where the
".config" file was copied from a source with the same path
as the destination, hence leading to a crash.

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>