config: Add 'workers' setting to config file
Add a new setting to the configuration file for specifying
the maximum number of threads/workers that klp-build shall
use in order to parallelize some tasks.
```
[Settings]
workers = 16
```
The '--workers' commandline option will be deprecated in
favor of the new setting. For the "extract" command, the
default value will remain as '4' for backward compatibility.

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>