ibs: push: Get the codestream branch and rebase it on top of base branch
Until now it was required to the developer to always base the branch on top of one of
branches assigned to the codestreams. As the number of base branches increase, it's easier
if klp-build could make it automatically.

From now on, it's expected that the bscXXXXXX branch is based on the kgraft-patches' master
branch, which doesn't include any code, since klp-build will clone that branch, and rebase it on
top of a base branch related to the codestream affected.

For example, for 12.5 codestreams, it will get the branch for bscXXXXX_12.5uXXX (which should contain
only the directory for the bscXXX) and rebase on top of origin/master-livepatch-sle12.

Hopefully this change will make the LP developers to check less details when creating a livepatch.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>