klplib: codestream: drop data_path, lp_path and lp_name fields
These fields are not needed for two reasons:

1.) We do not need to store this field in the codestream.json file as
they can be retrieved from the info stored in the config file (the path
to the work_dir and to data_dir) and the lp_name.

2.) Having to instantiate eveytime a `Codestream` object using these
fields, force us to use them even when not needed. This is one of the
reasons why we are currently using fake lp_name even when the function
in question does not need any.

So drop these fields to allow the scan command to work even without a
given lp_name (bsc_check).

Signed-off-by: Vincenzo MEZZELA <vincenzo.mezzela@suse.com>