klplib: cmd: reorganize argument parser
Previously, two distinct argument parsers were used: parentparser and
parser.

- 'parentparser' defined the --name and --filter arguments.
- 'parser' was primarily used to create sub, which handled subparser
  registration for individual commands.
- Commands requiring --name and --filter used 'parentparser' as a parent.

This patch merge the two parent parsers into a single one. Instead of
passing 'parentparser' to subparsers, we now manually register the
--name and --filter arguments only in the commands that require them.

While this change might seem like a step backward in complexity, it is
necessary to decouple the plugin subparser management from
'klplib/cmd.py'. Keeping plugin-related logic outside of this package
allows each plugin to handle its own subparser registration. This makes
plugins self-contained and eliminates the need to modify internal klplib
code when working solely on a plugin.

Additionally, this change clarifies the role of the parent parser, which
should now be used exclusively for general arguments. For example, in
the future, we could introduce a global --verbose option:

	klp-build --verbose scan --cve XXXX-XXXX

Signed-off-by: Vincenzo MEZZELA <vincenzo.mezzela@suse.com>