setup: Add a more advanced versioning schema
We are currently using the release `tag` as the only versioning
schema; meaning that all commits within the same release share
the same version. This obviously poses a limitation for developers
working with several builds and who want to keep track of local
changes. As such, I propose using the following versioning schema:

For releases: {tag}
For each commit after release: {tag}.post{commit-position}+git.{commit-hash}
For a commit with unstaged changes: {same-as-above}.dirty

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>