plugins: extractor: Group files based on generated livepatches
Until this commit the groupings were done using the files generated from klp-ccp,
meaning that code still had to include the templates. By comparing the files with
the final LP ones makes the grouping more accurate, meaning that we can rely more on it.

The template generation generates different code for different codestreams based on
specifics of each kernel version. And since the kgraft-patches codestream specific branches
are not used anymore, this means that we now have a more reliable way of determing how
many versions of each livepatch we need.

Closes: #72

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>