codestreams.py: Detect if a .ko module is missing
When clang-extract was being used we managed to pass a .gz or a .zst file to. But
klp-ccp is not able to do that, so we need to make sure that we find a .ko.

This should not be a problem since we download the rpms and then we extract the
the final .ko files, but if one stops the process it might be leftover, so inform which
file was not found.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>