klplib: plugins: use underscore in plugin file name
When implementing a new command, we might want to have "-" in the name,
see for instance the "format-patches" command. Unfortunately, we cannot
have the "-" in a python file name. Hence we need to replace the symbol
in the plugin loading machinery with "_" and then be consistent with the
plugin file naming: "format-patches" command will have its own plugin
named "format_patches.py" in the plugins package.

Signed-off-by: Vincenzo MEZZELA <vincenzo.mezzela@suse.com>