extract: Get gcc args by using compile_commands.json
Our make strategy worked well until we found the a new way of getting
the gcc arguments used to compile a kernel source file. Recently it started
to fail on newer codestreams, failing to find the proper _source_ directory from the
-obj directory.

Use this opportunity to finally implement the arguments finding using compile_commands.json
that is packaged by kernel-livepatch-default-livepatch-devel rpm.

The paths are extactly the same being found using make, but the instead of the
absolute paths to the source directory of the codestreams, the compile_commands.json
used '../'. Replacing ../ by the source directory of the affected codestream fixes
the problem.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>