setup: Make validate_config to check for specified archs
Up until now we could have a CONFIG_ being set on an archtecture and not
on others and then it would pass the checks, only complaining for a missing
symbol later on the setup phase.

IIRC this used to work in the past, but somehow it got broken. Added a test
to ensure that we catch such cases in the future.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>