scan: Ignore code streams which do not match --filter
The packages for all code streams are currently downloaded and
decompressed even when the caller is interested only into
a particular one.

As a result, "klp-build setup" command runs much longer it could.
It is not good in production and it makes debugging really painful.

The missing data are collected in scan(). Ignore the filtered
code streams there.

Signed-off-by: Petr Mladek <pmladek@suse.com>