klplib: codestream: use ksrc to read config
This change enable the possibility to read the config without relying on
any previously downloaded rpms. This means that the scan command can now
perform a more in deep analysis without consuming disk space with the
rpms.

The content of the config in the kernel-source git directory slightly
differs from the one in the rpms, this is likely due to the building
phase modifying some entries related to the compiler. Nevertheless, this
should not touch the entries that enable/disable portion of the source,
thus it shouldn't really matter.

Signed-off-by: Vincenzo Mezzela <vincenzo.mezzela@suse.com>