klplib/ksrc: Filter out merge commits
In some cases git-log incorrectly reports back the commit that
merged the branch containing the expected commit, and not the
expected commit itself. This happens even with the "--no-merge"
flag enabled, so there's no choice but to skip them manually.

Also, in order to find the commit that introduced the patch,
from now on git-log will be forced to follow all the parent commits
and traverse the full patch's history, until file creation.
This is a more exhaustive method, and as such slower than
before (~4 second slowdown).

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>