klplib/ksrc: Improve commit heuristic for RT kernel
Usually the commit intruducing the patch is ported from the non-RT
branch to the RT one.
If for whatever reason we fail to find any CVE-related commit in the RT branch,
but we do in the non-RT branch, try re-using the patch files found in non-RT.
If those files do not exist in the RT branch, then we can assume RT is
indeed not affected.

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>