klplib/ksrc: Simplify heuristic design
The current design for finding whether a kernel is patched or not
is extremely slow and inefficient. It relies heavily on git commits,
which inevitably requires processing the full log history of each
kernel branch/tag.

A simpler yet still effective design would be to just check if
the kernel branch/tag has the patch file in ./patches.suse/.

There might be some corner cases, such as patch renaming, but
it would be trivial to workaround those issues. What's more,
this approach is quite fast and requires much less code.

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>