klplib: ibs: Remove cs_data dictionary
In the past there were more elements to the cs_data dict, but it was
removed in another refactor. Right now we get the package name
from the Codestream object, and we only filter the rpms based on the regex.

Removing the dict also simplifies the flow of the function.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>