plugins/setup: Automate setup with scan report
From now on, setup will no longer require more parameters other than
'--cve' and '--name'. The livepatch technical information (that is: files,
functions, configs, modules and archs) will be directly retrieved from
the report generated by scan(). The user will only need to double-check
that the livepatch information is actually correct. In case it is not,
there's still as a fallback option the manual input ('--conf', '--file-func',
etc).

In addition, most of the setup functions have been simplified and reworked,
and the already existing tests have been adapted to the new changes.

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>