plugins/extract: Apply the patches with git
The current implementation uses quilt to apply the patches,
and although it works in most cases, it lacks the flexibility
offered by git. In the past it made sense to use quilt due to
several limitations in the tooling, but that's not the case anymore.
Keep patch(1) as a fallback in case git-am(1) fails.

From now on, during 'extract' each codestream will have its own
working branch, making it easier to track changes in case of errors.
From a high level the process will remain the same as before:
0. run extract()
1. remove patches
2. apply patches
3. run klp-ccp
4. remove patches

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>