plugins/commit: Add new 'commit' command
Provide a new command for commiting the extracted livepatches
to the kgraft-patches repository. The branches are created as
per indicated in the ccp/groups file. The default policy forbids
from modifying already existing branches, to avoid erasing any
user-changes. A "--force" option is provided to overwrite them.

A basic use-case would be as follow:
$ klp-build extract -n bsctest
$ klp-build commit -n bsctest
... # do some changes to the livepatches
$ klp-build push -n bsctest

Should the user need to extract again the livepatch, a 'force'
would be required:
$ klp-build extract -n bsctest
$ klp-build commit -n bsctest --force
...
$ klp-build push -n bsctest

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>