codestream: Add check for symbol being patchable
Kernel code can add instructions to the compiler do not allow the symbol
to be traceable/profiled, which also disabled the livepatching of the symbol.

Such cases can be tricky to found, and usually discovered only in the test
phase of a livepatch development. This change introduces a way to detect
such symbols by looking at the ftrace addresses created on build time.

For now it only checks the symbols on vmlinux. To extend the same solution
to kernel modules would involve resolving relocation manually, which is tricky.

Also, ppc64le is different regarding the content of start_mcount_loc and
stop_mcount_loc, since it gets an address pointing to function descriptors,
so the code code checks for the range of the symbol + symbol size.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>