klplib: ibs: Always extract all rpm packages for all architectures
If a given livepatch only affects a non x86 architecture the klp-build
tool will fail to find the ipa-clones files and other files that are
present on kernel-default-devel. Since the biggest rpm was kernel-source,
and we are not downloading and extracting it anymore, just extract
everything else like we do for x86.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>