file2config: Check if a given file is only compiled on one arch
There is a corner case of when a file being patched lives inside arch,
it would wrongly says that all other architectures are affected, for
example, arch/s390/kernel/perf_cpum_sf.c. The current code will find
the CONFIG_PERF_EVENTS entry, and tell that all archs are affected, when
in fact only s390x is.

This is not the complete solution, given that a specific CONFIG_$ARCH
must be checked with the additional config found, but we may extend this
functionality when such corner case is found.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>