templ: Use cs.archs when checking if IS_ENABLED should be used
The current check doesn't work on codestreams that are enabled on only
a subset of the supported architectures, like MICRO (which is enabled
only on x86 and s390x).

This happens because the function checks for all ARCHS against the affected
ones. To fix it compare the affected architectures with the architectures
supported on the codestream.

Also, sort the entries of the __default_archs method since it's used
on Templ class to compare with the affected one.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>