templ: Check get_default_archs with cs.archs
The original check was working by change. Each codestream has a set
of supported architectures, like -RT ones being supported only on x86_64
for now. The current check was checking the supported archs on a
codestream with the architectures that was affected on the project level,
which is wrong.

The change checks the affected architectures on the codestream against
the architectures that the codestream supports. The original tests also
worked with this change, being even more clear that it worked by chance
at the time.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>