plugins: setup: Change the checks for missing symbol
Up until now, whenever a symbol can't be found on an arch a warning was
shown. This isn't really a problem anymore since klp-ccp can pick the
callers of inline functions automatically, so it just adds more concern
to livepatch developers.

Change the procedure to check if the symbol is missing on all kernels,
which means that the same inline decision was made for all of them. But
if the symbol is present in one arch and not in another, this means that
maybe we can have issues, so print the warning as before.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>